use our AES API in Kyber
