+++ crypto_kem/kyber768-90s/m4/symmetric-aes.h
@@ -3,10 +3,11 @@
 
 #include <stddef.h>
 #include <stdint.h>
-
+#include "aes.h"
 typedef struct {
-    uint64_t sk_exp[120];
-    uint32_t ivw[16];
+    aes256ctx sk_exp;
+    uint8_t iv[12];
+    uint32_t ctr;
 } aes256xof_ctx;
 
 void aes256_prf(uint8_t *output, size_t outlen, const uint8_t *key, uint8_t nonce);
