+++ crypto_kem/kyber768-90s/m4/symmetric.h
@@ -4,7 +4,7 @@
 #include "params.h"
 
 
-#include "aes256ctr.h"
+#include "symmetric-aes.h"
 #include "sha2.h"
 
 #if (KYBER_SSBYTES != 32)
